+++ packages/bluesky/CHANGELOG.md
@@ -5,7 +5,7 @@
 - Added `cursor` to `Repos` object. ([#627](https://github.com/myConsciousness/atproto.dart/issues/627))
 - Upgraded `atproto_core` and exposed `RateLimit` and `RateLimitPolicy` objects. ([#636](https://github.com/myConsciousness/atproto.dart/issues/636))
 - Added `langs` field in `PostRecord`. ([#642](https://github.com/myConsciousness/atproto.dart/issues/642))
-- Added pagination utils.
+- Added pagination utils. ([#626](https://github.com/myConsciousness/atproto.dart/issues/626))
   - `ActorsService`
     - `paginateActors`
     - `paginateActorsAsJson`